Further tests written for Odometer adjustment rides.
Tests written & working for adjustment rides involving mixed distance
units (km & miles),
and deletion/creation of adjustment ride when
Odometer.initial is changed

Code & tests not yet written for automatic calling of
update_adjustment_ride when an Odometer is saved, or when a Ride between
two Odometer readings is saved.

And nothing in a view calls update_adjustment_rides yet, so it is never
invoked, but it is fully working & tested.